Implemented notes replacement
