+++ packages/stencil-library/package.json
@@ -45,6 +45,7 @@
     "@types/puppeteer": "5.4.7",
     "@typescript-eslint/eslint-plugin": "^4.33.0",
     "@typescript-eslint/parser": "^4.33.0",
+    "babel-loader": "^9.1.2",
     "eslint": "^7.32.0",
     "eslint-plugin-react": "^7.16.0",
     "eslint-plugin-storybook": "^0.6.11",
